chore(api): Phase 3 - Complete module migration
- Created modular API structure with 12 domain files:
  * teams.ts (7 functions)
  * games.ts (5 functions)
  * locations.ts (2 functions)
  * membership.ts (12 functions)
  * practices.ts (3 functions)
  * barber.ts (4 functions)
  * events.ts (2 functions)
  * users.ts (3 functions + Waiver type)

- Simplified utils/api.ts to single re-export line
- Updated utils/api/index.ts with organized exports
- Enhanced verification scripts:
  * Handle multi-line export syntax
  * Support wildcard re-exports
  * Improved comparison logic

Verification:
✅ TypeScript compilation passes
✅ All 49 exports available (45 named + 4 via wildcards)
✅ Zero breaking changes
