+++ utils/api.ts
@@ -1,2043 +1,2 @@
-import axios from "axios";
-import { USE_MEMBERSHIP_TEST_MODE, MEMBERSHIP_TEST_PLAN_ID } from "./constants";
-import { createUserWithEmailAndPassword, signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from "@/firebase/firebaseConfig";
-import AsyncStorage from "@react-native-async-storage/async-storage";
-
-// ⚠️ MIGRATED TO: utils/api/core/constants.ts
-// export const API_URL = "https://api-461776259687.us-west2.run.app";
-
-// ⚠️ MIGRATED TO: utils/api/core/types.ts
-// export type APIErrorType = ...
-// export interface APIError { ... }
-// export interface APIResponse<T> { ... }
-// type User = { ... }
-
-// ⚠️ MIGRATED TO: utils/api/core/client.ts
-// export const refreshBackendJwt = async (): Promise<string> => { ... }
-
-// Import from new structure for internal use
-import { API_URL } from './api/core/constants';
-import { type APIErrorType, type APIError, type APIResponse, type User } from './api/core/types';
-import { refreshBackendJwt as _refreshBackendJwt } from './api/core/client';
-import {
-  loginUser as _loginUser,
-  verifyEmail as _verifyEmail,
-  resendVerificationEmail as _resendVerificationEmail,
-  registerChild as _registerChild,
-  registerUser as _registerUser,
-} from './api/auth';
-
-// Re-export for external consumers (temporary bridge)
-export { API_URL };
-export type { APIErrorType, APIError, APIResponse, User };
-export const refreshBackendJwt = _refreshBackendJwt;
-
-// Phase 2: Auth exports
-export const loginUser = _loginUser;
-export const verifyEmail = _verifyEmail;
-export const resendVerificationEmail = _resendVerificationEmail;
-export const registerChild = _registerChild;
-export const registerUser = _registerUser;
-
-// (Removed global axios interceptor to avoid global side-effects)
-
-
-// ⚠️ MIGRATED TO: utils/api/auth.ts
-// - loginUser
-// - verifyEmail
-// - resendVerificationEmail
-// - registerChild
-// - registerUser
-
-
-
-// 🔹 **Team API Functions**
-
-// Get all teams for the authenticated coach
-export const getTeams = async (token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-    
-    // Try to get Firebase token if user is available, but don't require it
-    // Since testing shows this endpoint works with just JWT token
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-      }
-    } else {
-    }
-    
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-    
-    // Add Firebase token if available
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-    
-    const response = await axios.get(`${API_URL}/secure/teams`, {
-      headers,
-    });
-    
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to fetch teams:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Get team details by ID including roster
-export const getTeamById = async (teamId: string, token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    // Try to get Firebase token if user is available, but don't require it
-    // Since testing shows this endpoint works with just JWT token
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-      }
-    } else {
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    // Add Firebase token if available
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.get(`${API_URL}/teams/${teamId}`, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to fetch team ${teamId}:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Create a new team
-export const createTeam = async (teamData: { name: string; capacity: number; logo?: string; coach_id?: string }, token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.post(`${API_URL}/teams`, teamData, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to create team:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Update an existing team
-export const updateTeam = async (teamId: string, teamData: { name: string; capacity: number; coach_id: string; logo?: string }, token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.put(`${API_URL}/teams/${teamId}`, teamData, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to update team ${teamId}:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Delete a team
-export const deleteTeam = async (teamId: string, token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.delete(`${API_URL}/teams/${teamId}`, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to delete team ${teamId}:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// 🔹 **External Team API Functions**
-
-// Get all external teams (opponent teams)
-export const getExternalTeams = async (): Promise<any> => {
-  try {
-    const response = await axios.get(`${API_URL}/teams/external`);
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to fetch external teams:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Create a new external team
-export const createExternalTeam = async (
-  teamData: { name: string; capacity: number; logo_url?: string },
-  token: string
-): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.post(`${API_URL}/teams/external`, teamData, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to create external team:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// 🔹 **Location API Functions**
-
-// Get all locations
-export const getLocations = async (): Promise<any> => {
-  try {
-    const response = await axios.get(`${API_URL}/locations`);
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to fetch locations:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Get location details by ID
-export const getLocationById = async (locationId: string): Promise<any> => {
-  try {
-    const response = await axios.get(`${API_URL}/locations/${locationId}`);
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to fetch location ${locationId}:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// 🔹 **Game API Functions**
-
-// Get all games for the authenticated coach
-export const getGames = async (token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    // Try to get Firebase token if user is available
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    // Add Firebase token if available
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.get(`${API_URL}/secure/games`, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to fetch games:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Get game details by ID
-export const getGameById = async (gameId: string, token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.get(`${API_URL}/games/${gameId}`, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to fetch game ${gameId}:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Create a new game
-export const createGame = async (
-  gameData: {
-    home_team_id: string;
-    away_team_id: string;
-    location_id: string;
-    start_time: string;
-    court_id?: string;
-    end_time?: string;
-    status?: string;
-    home_score?: number;
-    away_score?: number;
-  },
-  token: string
-): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.post(`${API_URL}/games`, gameData, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to create game:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Update an existing game
-export const updateGame = async (
-  gameId: string,
-  gameData: {
-    home_team_id?: string;
-    away_team_id?: string;
-    location_id?: string;
-    start_time?: string;
-    court_id?: string;
-    end_time?: string;
-    status?: string;
-    home_score?: number;
-    away_score?: number;
-  },
-  token: string
-): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.put(`${API_URL}/games/${gameId}`, gameData, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to update game ${gameId}:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Delete a game
-export const deleteGame = async (gameId: string, token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.delete(`${API_URL}/games/${gameId}`, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to delete game ${gameId}:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// 🔹 **Membership API Functions**
-
-// Get all available credit packages (no authentication required)
-export const getCreditPackages = async () => {
-  try {
-    const response = await fetch(`${API_URL}/credit_packages`, {
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      return {
-        data: null,
-        error: {
-          message: `Failed to fetch credit packages: ${response.status} ${errorText}`,
-          status: response.status
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to fetch credit packages:", error);
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to fetch credit packages",
-        status: 500
-      }
-    };
-  }
-};
-
-// Get all available membership plans (no authentication required)
-export const getAllMembershipPlans = async () => {
-  try {
-    const response = await fetch(`${API_URL}/memberships`, {
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      return {
-        data: null,
-        error: {
-          message: `Failed to fetch membership plans: ${response.status} ${errorText}`,
-          status: response.status
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to fetch membership plans:", error);
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to fetch membership plans",
-        status: 500
-      }
-    };
-  }
-};
-
-// Get current user's membership details (requires authentication)
-export const getUserMemberships = async (): Promise<APIResponse<any>> => {
-  try {
-    // Get stored JWT token for backend authentication
-    // The project uses "Firebase for identity, JWT for business authorization"
-    let jwtToken = await AsyncStorage.getItem("authToken");
-
-    // If no JWT token, we need to get one from /auth endpoint using Firebase token
-    if (!jwtToken) {
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.error("❌ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401,
-            type: 'auth',
-            userMessage: "Session expired. Please log in again."
-          }
-        };
-      }
-    }
-
-
-    // Use the correct endpoint for current logged-in user: /secure/customers/memberships
-    const response = await fetch(`${API_URL}/secure/customers/memberships`, {
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-      },
-    });
-
-    if (!response.ok) {
-      const errorText = await response.text();
-
-      // If 401/403, try refreshing the JWT token once
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-
-          // Retry the request with new token
-          const retryResponse = await fetch(`${API_URL}/secure/customers/memberships`, {
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-            },
-          });
-
-          if (!retryResponse.ok) {
-            const retryErrorText = await retryResponse.text();
-
-            // Classify retry error (aligned with backend)
-            const retryErrorType: APIErrorType =
-              retryResponse.status === 401 ? 'auth' :
-              retryResponse.status === 403 ? 'permission' :
-              retryResponse.status === 404 ? 'not_found' :
-              retryResponse.status === 408 ? 'timeout' :
-              retryResponse.status === 409 ? 'conflict' :
-              retryResponse.status === 410 ? 'gone' :
-              retryResponse.status === 413 ? 'payload_too_large' :
-              retryResponse.status === 429 ? 'rate_limit' :
-              retryResponse.status === 503 ? 'service_unavailable' :
-              retryResponse.status >= 500 ? 'server' :
-              retryResponse.status === 400 ? 'validation' :
-              'unknown';
-
-            const retryUserMessage =
-              retryResponse.status === 401 ? "Session expired. Please log in again." :
-              retryResponse.status === 403 ? "You don't have permission to view membership information." :
-              retryResponse.status === 404 ? "Membership information not found." :
-              retryResponse.status === 408 ? "Request timeout. Please try again." :
-              retryResponse.status === 409 ? "This action conflicts with existing data. Please refresh and try again." :
-              retryResponse.status === 410 ? "This resource is no longer available. Please contact support." :
-              retryResponse.status === 413 ? "File size too large. Maximum 10MB allowed." :
-              retryResponse.status === 429 ? "Too many requests. Please wait a moment and try again." :
-              retryResponse.status === 503 ? "Service is temporarily unavailable. Please try again later." :
-              retryResponse.status >= 500 ? "Server error. Please try again later." :
-              retryResponse.status === 400 ? "Invalid request. Please contact support." :
-              "Unable to load membership information. Please try again.";
-
-            return {
-              data: null,
-              error: {
-                message: `Failed to fetch user memberships after token refresh: ${retryResponse.status} ${retryErrorText}`,
-                status: retryResponse.status,
-                type: retryErrorType,
-                userMessage: retryUserMessage
-              }
-            };
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (refreshError) {
-          console.error("❌ Failed to refresh token on retry:", refreshError);
-          return {
-            data: null,
-            error: {
-              message: `Authentication failed: ${response.status} ${errorText}`,
-              status: response.status,
-              type: 'auth',
-              userMessage: "Session expired. Please log in again."
-            }
-          };
-        }
-      }
-
-      // Classify error based on status code (aligned with backend)
-      const errorType: APIErrorType =
-        response.status === 401 ? 'auth' :
-        response.status === 403 ? 'permission' :
-        response.status === 404 ? 'not_found' :
-        response.status === 408 ? 'timeout' :
-        response.status === 409 ? 'conflict' :
-        response.status === 410 ? 'gone' :
-        response.status === 413 ? 'payload_too_large' :
-        response.status === 429 ? 'rate_limit' :
-        response.status === 503 ? 'service_unavailable' :
-        response.status >= 500 ? 'server' :
-        response.status === 400 ? 'validation' :
-        'unknown';
-
-      const userMessage =
-        response.status === 401 ? "Session expired. Please log in again." :
-        response.status === 403 ? "You don't have permission to view membership information." :
-        response.status === 404 ? "Membership information not found." :
-        response.status === 408 ? "Request timeout. Please try again." :
-        response.status === 409 ? "This action conflicts with existing data. Please refresh and try again." :
-        response.status === 410 ? "This resource is no longer available. Please contact support." :
-        response.status === 413 ? "File size too large. Maximum 10MB allowed." :
-        response.status === 429 ? "Too many requests. Please wait a moment and try again." :
-        response.status === 503 ? "Service is temporarily unavailable. Please try again later." :
-        response.status >= 500 ? "Server error. Please try again later." :
-        response.status === 400 ? "Invalid request. Please contact support." :
-        "Unable to load membership information. Please try again.";
-
-      return {
-        data: null,
-        error: {
-          message: `Failed to fetch user memberships: ${response.status} ${errorText}`,
-          status: response.status,
-          type: errorType,
-          userMessage: userMessage
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to fetch user memberships:", error);
-
-    // Check if it's a network error
-    const isNetworkError =
-      error instanceof TypeError &&
-      (error.message.includes('Network request failed') ||
-       error.message.includes('Failed to fetch') ||
-       error.message.includes('timeout'));
-
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to fetch user memberships",
-        status: 500,
-        type: isNetworkError ? 'network' : 'unknown',
-        userMessage: isNetworkError
-          ? "Please check your internet connection and try again."
-          : "An unexpected error occurred. Please try again."
-      }
-    };
-  }
-};
-
-// Get all available membership plans (requires authentication)
-export const getMembershipPlans = async () => {
-  try {
-    // Get stored JWT token for backend authentication
-    // The project uses "Firebase for identity, JWT for business authorization"
-    let jwtToken = await AsyncStorage.getItem("authToken");
-
-    // If no JWT token, we need to get one from /auth endpoint using Firebase token
-    if (!jwtToken) {
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.error("❌ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401,
-            type: 'auth'
-          }
-        };
-      }
-    }
-
-    // Use the correct endpoint: /memberships (as confirmed by test results)
-    const response = await fetch(`${API_URL}/memberships`, {
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-        "Content-Type": "application/json",
-      },
-    });
-
-    if (!response.ok) {
-      const errorText = await response.text();
-
-      // If 401/403, try refreshing the JWT token once
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-
-          // Retry the request with new token
-          const retryResponse = await fetch(`${API_URL}/memberships`, {
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-              "Content-Type": "application/json",
-            },
-          });
-
-          if (!retryResponse.ok) {
-            return {
-              data: null,
-              error: {
-                message: `Failed to fetch membership plans: ${retryResponse.status} ${errorText}`,
-                status: retryResponse.status
-              }
-            };
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (retryError) {
-          console.error("❌ Failed to refresh JWT token on retry:", retryError);
-          return {
-            data: null,
-            error: {
-              message: "Unable to authenticate with backend",
-              status: 401
-            }
-          };
-        }
-      }
-
-      return {
-        data: null,
-        error: {
-          message: `Failed to fetch membership plans: ${response.status} ${errorText}`,
-          status: response.status
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to fetch membership plans:", error);
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to fetch membership plans",
-        status: 500
-      }
-    };
-  }
-};
-
-
-// Get specific pricing plans for a membership type (requires authentication)
-export const getPlansForMembership = async (membershipId: string, jwtOverride?: string) => {
-  try {
-    // Get stored JWT token for backend authentication
-    // The project uses "Firebase for identity, JWT for business authorization"
-    let jwtToken = jwtOverride || (await AsyncStorage.getItem("authToken"));
-
-    // If no JWT token, we need to get one from /auth endpoint using Firebase token
-    if (!jwtToken) {
-      if (jwtOverride) {
-        return {
-          data: null,
-          error: {
-            message: "Missing authentication token",
-            status: 401,
-            type: 'auth'
-          }
-        }
-      }
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.error("❌ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401,
-            type: 'auth'
-          }
-        };
-      }
-    }
-
-    // Use the endpoint: /memberships/{membershipId}/plans
-    const requestUrl = `${API_URL}/memberships/${membershipId}/plans`;
-    const headers = {
-      "Authorization": `Bearer ${jwtToken}`,
-    };
-
-    const response = await fetch(requestUrl, {
-      headers,
-    });
-
-    const handleErrorResponse = async (resp: Response) => {
-      const errorText = await resp.text();
-      const baseError = {
-        message: `Failed to fetch plans for membership ${membershipId}: ${resp.status} ${errorText}`,
-        status: resp.status,
-        type: resp.status === 401 || resp.status === 403 ? 'auth' : 'api'
-      } as const
-
-      // Don't log 503 errors (known issue with incomplete membership configurations)
-      if (resp.status !== 503) {
-        console.error(`❌ Request failed for membership ${membershipId}:`, {
-          url: requestUrl,
-          status: resp.status,
-          statusText: resp.statusText,
-          errorBody: errorText
-        });
-      }
-
-      return baseError
-    }
-
-    if (!response.ok) {
-      const errorText = await response.text();
-
-      // Don't log 503 errors (known issue with incomplete membership configurations)
-      if (response.status !== 503) {
-        console.error(`❌ Request failed for membership ${membershipId}:`, {
-          url: requestUrl,
-          status: response.status,
-          statusText: response.statusText,
-          errorBody: errorText
-        });
-      }
-
-      // If 401/403, try refreshing the JWT token once
-      if ((response.status === 401 || response.status === 403) && !jwtOverride) {
-        try {
-          jwtToken = await refreshBackendJwt();
-
-          const retryResponse = await fetch(`${API_URL}/memberships/${membershipId}/plans`, {
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-            },
-          });
-
-          if (!retryResponse.ok) {
-            const retryError = await handleErrorResponse(retryResponse)
-            return { data: null, error: retryError }
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (retryError) {
-          console.error("❌ Failed to refresh JWT token on retry:", retryError);
-          return {
-            data: null,
-            error: {
-              message: "Unable to authenticate with backend after retry",
-              status: 401,
-              type: 'auth'
-            }
-          };
-        }
-      }
-
-      return {
-        data: null,
-        error: await handleErrorResponse(response)
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error(`❌ Unexpected error while fetching plans for membership ${membershipId}:`, error);
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || `Failed to fetch plans for membership ${membershipId}`,
-        status: 500,
-        type: 'network'
-      }
-    };
-  }
-};
-
-// Initiate membership plan purchase (requires authentication)
-export const purchaseMembershipPlan = async (planId: string) => {
-  console.log("🔵 purchaseMembershipPlan called with planId:", planId);
-  try {
-    // Get stored JWT token for backend authentication
-    // The project uses "Firebase for identity, JWT for business authorization"
-    let jwtToken = await AsyncStorage.getItem("authToken");
-    console.log("🔵 JWT token retrieved:", jwtToken ? "✅ exists" : "❌ missing");
-
-    // If no JWT token, we need to get one from /auth endpoint using Firebase token
-    if (!jwtToken) {
-      console.log("🔵 No JWT token found, attempting to refresh...");
-      try {
-        jwtToken = await refreshBackendJwt();
-        console.log("🔵 JWT token refreshed successfully");
-      } catch (refreshError) {
-        console.warn("⚠️ Failed to refresh JWT token:", refreshError);
-        return { data: null, error: { status: 401, message: "Unable to authenticate with backend" } } as any;
-      }
-    }
-
-    // Use test mode flag to determine effective plan id
-    const effectivePlanId = USE_MEMBERSHIP_TEST_MODE ? MEMBERSHIP_TEST_PLAN_ID : planId;
-    console.log("🔵 Test mode:", USE_MEMBERSHIP_TEST_MODE, "| Effective plan ID:", effectivePlanId);
-
-    const url = `${API_URL}/checkout/membership_plans/${effectivePlanId}`;
-    console.log("🔵 Making POST request to:", url);
-
-    const response = await fetch(url, {
-      method: "POST",
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-        "Content-Type": "application/json",
-      },
-    });
-
-    console.log("🔵 Response status:", response.status, response.statusText);
-
-    if (!response.ok) {
-      // Try to read structured error from backend
-      let errorText = await response.text();
-      console.log("🔵 Error response body:", errorText);
-      let errorMessage = "Failed to initiate membership purchase";
-      try {
-        const parsed = JSON.parse(errorText || "{}");
-        errorMessage = parsed?.error?.message || errorMessage;
-      } catch {}
-
-      // If 401/403, try refreshing the JWT token once
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-
-          // Retry the request with new token
-          const retryResponse = await fetch(`${API_URL}/checkout/membership_plans/${effectivePlanId}`, {
-            method: "POST",
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-              "Content-Type": "application/json",
-            },
-          });
-
-          if (!retryResponse.ok) {
-            const retryErrorText = await retryResponse.text();
-            let retryMessage = errorMessage;
-            try {
-              const parsed = JSON.parse(retryErrorText || "{}");
-              retryMessage = parsed?.error?.message || retryMessage;
-            } catch {}
-            // Localize 409 handling on retry path
-            if (retryResponse.status === 409) {
-              return { data: null, error: { status: 409, message: retryMessage } } as any;
-            }
-
-            // Return structured error instead of throwing
-            return { data: null, error: { status: retryResponse.status, message: retryMessage } } as any;
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null } as any;
-        } catch (refreshError) {
-          console.warn("⚠️ Failed to refresh token on retry:", refreshError);
-          return { data: null, error: { status: response.status, message: `Authentication failed: ${errorMessage}` } } as any;
-        }
-      }
-
-      // Localize 409 handling: return structured error object instead of throwing
-      if (response.status === 409) {
-        return { data: null, error: { status: 409, message: errorMessage } } as any;
-      }
-
-      // Return structured error instead of throwing
-      return { data: null, error: { status: response.status, message: errorMessage } } as any;
-    }
-
-    const okData = await response.json();
-    console.log("✅ Purchase successful! Response data:", okData);
-    return { data: okData, error: null } as any;
-  } catch (error) {
-    console.error("❌ Failed to purchase membership plan - Exception caught:", error);
-    const anyErr: any = error;
-    if (anyErr?.status === 409) {
-      return { data: null, error: { status: 409, message: (error as Error).message } } as any;
-    }
-    return { data: null, error: { status: anyErr?.status || 500, message: (error as Error).message } } as any;
-  }
-};
-
-// Initiate credit package purchase (requires authentication)
-export const purchaseCreditPackage = async (packageId: string) => {
-  try {
-    // Get stored JWT token for backend authentication
-    let jwtToken = await AsyncStorage.getItem("authToken");
-
-    // If no JWT token, we need to get one from /auth endpoint using Firebase token
-    if (!jwtToken) {
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.warn("⚠️ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401
-          }
-        };
-      }
-    }
-
-    const response = await fetch(`${API_URL}/checkout/credit_packages/${packageId}`, {
-      method: "POST",
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-        "Content-Type": "application/json",
-      },
-    });
-
-    if (!response.ok) {
-      // Try to read structured error from backend
-      let errorText = await response.text();
-      let errorMessage = "Failed to initiate credit package purchase";
-      try {
-        const parsed = JSON.parse(errorText || "{}");
-        errorMessage = parsed?.error?.message || errorMessage;
-      } catch {}
-
-      // If 401/403, try refreshing the JWT token once
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-
-          // Retry the request with new token
-          const retryResponse = await fetch(`${API_URL}/checkout/credit_packages/${packageId}`, {
-            method: "POST",
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-              "Content-Type": "application/json",
-            },
-          });
-
-          if (!retryResponse.ok) {
-            const retryErrorText = await retryResponse.text();
-            let retryMessage = errorMessage;
-            try {
-              const parsed = JSON.parse(retryErrorText || "{}");
-              retryMessage = parsed?.error?.message || retryMessage;
-            } catch {}
-
-            return {
-              data: null,
-              error: {
-                status: retryResponse.status,
-                message: retryMessage
-              }
-            };
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (refreshError) {
-          console.warn("⚠️ Failed to refresh token on retry:", refreshError);
-          return {
-            data: null,
-            error: {
-              message: `Authentication failed: ${response.status} ${errorText}`,
-              status: response.status
-            }
-          };
-        }
-      }
-
-      return {
-        data: null,
-        error: {
-          status: response.status,
-          message: errorMessage
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.warn("⚠️ Failed to purchase credit package:", error);
-    return {
-      data: null,
-      error: {
-        status: 500,
-        message: (error as Error).message || "Failed to purchase credit package"
-      }
-    };
-  }
-};
-
-export const getMembershipByCustomerId = async (customerId: string) => {
-  // Get stored JWT token for backend authentication
-  let jwtToken = await AsyncStorage.getItem("authToken");
-
-  if (!jwtToken) {
-    try {
-      jwtToken = await refreshBackendJwt();
-    } catch (refreshError) {
-      console.error("❌ Failed to refresh JWT token:", refreshError);
-      return [];
-    }
-  }
-  const response = await fetch(`${API_URL}/customers/${customerId}/memberships`, {
-    headers: {
-      Authorization: `Bearer ${jwtToken}`,
-      "Content-Type": "application/json",
-    },
-  });
-
-  if (!response.ok) throw new Error("Failed to fetch customer membership");
-  return response.json();
-};
-
-export const createPractice = async (data: any, jwt: string) => {
-  const response = await fetch(`${API_URL}/events/one-time`, {
-    method: "POST",
-    headers: {
-      Authorization: `Bearer ${jwt}`,
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(data),
-  });
-
-  if (!response.ok) {
-    const err = await response.json();
-    throw new Error(err?.error?.message || "Failed to create practice");
-  }
-};
-
-export const createRecurringPractice = async (data: any, jwt: string) => {
-  const response = await fetch(`${API_URL}/events/recurring`, {
-    method: "POST",
-    headers: {
-      Authorization: `Bearer ${jwt}`,
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(data),
-  });
-
-  if (!response.ok) {
-    const err = await response.json();
-    throw new Error(err?.error?.message || "Failed to create recurring practice");
-  }
-};
-
-export const getPracticePrograms = async () => {
-  // Get stored JWT token for backend authentication
-  let jwtToken = await AsyncStorage.getItem("authToken");
-  
-  if (!jwtToken) {
-    try {
-      jwtToken = await refreshBackendJwt();
-    } catch (refreshError) {
-      console.error("❌ Failed to refresh JWT token:", refreshError);
-      throw new Error("User not logged in");
-    }
-  }
-
-  const res = await fetch(`${API_URL}/programs?program_type=practice`, {
-    headers: {
-      Authorization: `Bearer ${jwtToken}`,
-    },
-  });
-
-  if (!res.ok) {
-    const error = await res.json();
-    throw new Error(error?.message || "Failed to load programs");
-  }
-
-  return res.json(); // should return array of { id, name, ... }
-};
-
-// 🔹 **Haircut Booking API Functions**
-
-// Get all haircut services and barbers
-export const getHaircutAndBarberServices = async (): Promise<any> => {
-  try {
-    const response = await axios.get(`${API_URL}/haircuts/services`);
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to fetch haircut services:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Create a new haircut booking
-export const createHaircutBooking = async (bookingDetails: any, token: string): Promise<any> => {
-  try {
-    const response = await axios.post(`${API_URL}/haircuts/events`, bookingDetails, {
-      headers: {
-        "Authorization": `Bearer ${token}`,
-        "Content-Type": "application/json",
-      },
-    });
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to create haircut booking:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Get upcoming bookings for the authenticated user
-export const getUpcomingBookings = async (token: string): Promise<any> => {
-  try {
-    const response = await axios.get(`${API_URL}/bookings/upcoming`, {
-      headers: {
-        "Authorization": `Bearer ${token}`,
-      },
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to fetch upcoming bookings:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Get available time slots for a specific barber on a given date and for a given service duration
-export const getBarberAvailability = async (barberId: string, date: string, serviceDuration: number, token: string): Promise<string[]> => {
-  try {
-    const response = await axios.get(`${API_URL}/haircuts/barbers/${barberId}/availability`, {
-      headers: {
-        "Authorization": `Bearer ${token}`,
-      },
-      params: {
-        date,
-        service_duration: serviceDuration,
-      },
-    });
-    // Ensure we return the array from the 'available_slots' key, or an empty array as a fallback.
-    return response.data.available_slots || [];
-  } catch (error) {
-    console.error("❌ Failed to fetch barber availability:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// 🔹 **Credit API Functions**
-
-// Get current user's credit balance
-export const getUserCredits = async (token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.get(`${API_URL}/secure/credits`, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error("❌ Failed to fetch user credits:", (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Get event enrollment options including credit cost
-export const getEventEnrollmentOptions = async (eventId: string, token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.get(`${API_URL}/checkout/events/${eventId}/options`, {
-      headers,
-    });
-
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to fetch enrollment options for event ${eventId}:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// Enroll in event using credits
-export const enrollEventWithCredits = async (eventId: string, token: string): Promise<any> => {
-  try {
-    const firebaseUser = auth.currentUser;
-
-    let firebaseToken = null;
-    if (firebaseUser) {
-      try {
-        firebaseToken = await firebaseUser.getIdToken(true);
-      } catch (firebaseError) {
-        console.error("Failed to get Firebase token:", firebaseError);
-      }
-    }
-
-    const headers: Record<string, string> = {
-      "Authorization": `Bearer ${token}`,
-      "Content-Type": "application/json",
-    };
-
-    if (firebaseToken) {
-      headers["firebase_token"] = firebaseToken;
-    }
-
-    const response = await axios.post(
-      `${API_URL}/checkout/events/${eventId}/enhanced`,
-      { payment_method: "credits" },
-      { headers }
-    );
-
-    return response.data;
-  } catch (error) {
-    console.error(`❌ Failed to enroll in event ${eventId} with credits:`, (error as any).response?.data || (error as any).message);
-    throw error;
-  }
-};
-
-// 🔹 **Account Deletion API Function**
-
-// Delete user account
-export const deleteUserAccount = async (userToken: string) => {
-  try {
-    const response = await axios.delete(`${API_URL}/secure/customers/delete-account`, {
-      headers: {
-        'Authorization': `Bearer ${userToken}`,
-        'Content-Type': 'application/json',
-      },
-      data: {
-        confirm_deletion: true
-      }
-    });
-
-    if (response.status === 200) {
-      return { data: true, error: null };
-    } else {
-      return {
-        data: null,
-        error: {
-          message: `Account deletion failed with status: ${response.status}`,
-          status: response.status
-        }
-      };
-    }
-  } catch (error) {
-    console.error("❌ Account deletion API error:", error);
-    return {
-      data: null,
-      error: {
-        message: (error as any).response?.data?.error?.message || "Failed to delete account",
-        status: (error as any).response?.status || 500
-      }
-    };
-  }
-};
-
-// 🔹 **Subsidy API Functions**
-
-// Get user subsidy information
-export const getUserSubsidies = async (): Promise<APIResponse<any>> => {
-  try {
-    let jwtToken = await AsyncStorage.getItem("authToken");
-
-    if (!jwtToken) {
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.error("❌ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401,
-            type: 'auth',
-            userMessage: "Session expired. Please log in again."
-          }
-        };
-      }
-    }
-
-    const response = await fetch(`${API_URL}/subsidies/me`, {
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-      },
-    });
-
-    if (!response.ok) {
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-          const retryResponse = await fetch(`${API_URL}/subsidies/me`, {
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-            },
-          });
-
-          if (!retryResponse.ok) {
-            const retryErrorText = await retryResponse.text();
-            const retryErrorType: APIErrorType = retryResponse.status === 401 ? 'auth' : 'unknown';
-            const retryUserMessage = retryResponse.status === 401
-              ? "Session expired. Please log in again."
-              : "Unable to load subsidy information. Please try again.";
-
-            return {
-              data: null,
-              error: {
-                message: `Failed to fetch subsidies after token refresh: ${retryResponse.status} ${retryErrorText}`,
-                status: retryResponse.status,
-                type: retryErrorType,
-                userMessage: retryUserMessage
-              }
-            };
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (refreshError) {
-          return {
-            data: null,
-            error: {
-              message: "Token refresh failed",
-              status: 401,
-              type: 'auth',
-              userMessage: "Session expired. Please log in again."
-            }
-          };
-        }
-      }
-
-      const errorText = await response.text();
-      return {
-        data: null,
-        error: {
-          message: `Failed to fetch subsidies: ${response.status} ${errorText}`,
-          status: response.status,
-          type: 'unknown',
-          userMessage: "Unable to load subsidy information. Please try again."
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to fetch subsidies:", error);
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to fetch subsidies",
-        status: 500,
-        type: 'network',
-        userMessage: "Network error. Please check your connection and try again."
-      }
-    };
-  }
-};
-
-// Get user subsidy balance
-export const getUserSubsidyBalance = async (): Promise<APIResponse<any>> => {
-  try {
-    let jwtToken = await AsyncStorage.getItem("authToken");
-
-    if (!jwtToken) {
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.error("❌ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401,
-            type: 'auth',
-            userMessage: "Session expired. Please log in again."
-          }
-        };
-      }
-    }
-
-    const response = await fetch(`${API_URL}/subsidies/me/balance`, {
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-      },
-    });
-
-    if (!response.ok) {
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-          const retryResponse = await fetch(`${API_URL}/subsidies/me/balance`, {
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-            },
-          });
-
-          if (!retryResponse.ok) {
-            const retryErrorText = await retryResponse.text();
-            const retryErrorType: APIErrorType = retryResponse.status === 401 ? 'auth' : 'unknown';
-            const retryUserMessage = retryResponse.status === 401
-              ? "Session expired. Please log in again."
-              : "Unable to load subsidy balance. Please try again.";
-
-            return {
-              data: null,
-              error: {
-                message: `Failed to fetch subsidy balance after token refresh: ${retryResponse.status} ${retryErrorText}`,
-                status: retryResponse.status,
-                type: retryErrorType,
-                userMessage: retryUserMessage
-              }
-            };
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (refreshError) {
-          return {
-            data: null,
-            error: {
-              message: "Token refresh failed",
-              status: 401,
-              type: 'auth',
-              userMessage: "Session expired. Please log in again."
-            }
-          };
-        }
-      }
-
-      const errorText = await response.text();
-      return {
-        data: null,
-        error: {
-          message: `Failed to fetch subsidy balance: ${response.status} ${errorText}`,
-          status: response.status,
-          type: 'unknown',
-          userMessage: "Unable to load subsidy balance. Please try again."
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to fetch subsidy balance:", error);
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to fetch subsidy balance",
-        status: 500,
-        type: 'network',
-        userMessage: "Network error. Please check your connection and try again."
-      }
-    };
-  }
-};
-
-// 🔹 **Waiver API Functions**
-
-export interface Waiver {
-  id: string;
-  user_id?: string;
-  file_url: string;
-  file_name?: string;
-  file_size?: number;
-  file_type?: string;
-  notes?: string;
-  uploaded_at?: {
-    Time: string;
-    Valid: boolean;
-  };
-  uploaded_by?: string;
-}
-
-// Upload a waiver document
-export const uploadWaiver = async (
-  file: { uri: string; name: string; type: string },
-  userId: string,
-  notes?: string
-): Promise<APIResponse<Waiver>> => {
-  try {
-    let jwtToken = await AsyncStorage.getItem("authToken");
-
-    if (!jwtToken) {
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.error("❌ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401,
-            type: 'auth',
-            userMessage: "Session expired. Please log in again."
-          }
-        };
-      }
-    }
-
-    const formData = new FormData();
-    formData.append('file', {
-      uri: file.uri,
-      name: file.name,
-      type: file.type,
-    } as any);
-    formData.append('user_id', userId);
-
-    // Build URL with query parameters (notes only)
-    const params = new URLSearchParams();
-    if (notes) {
-      params.append('notes', notes);
-    }
-
-    // Build the URL with optional query string
-    const queryString = params.toString();
-    const uploadUrl = queryString ? `${API_URL}/waivers/upload?${queryString}` : `${API_URL}/waivers/upload`;
-
-    const response = await fetch(uploadUrl, {
-      method: 'POST',
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-      },
-      body: formData,
-    });
-
-    if (!response.ok) {
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-          const retryResponse = await fetch(uploadUrl, {
-            method: 'POST',
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-            },
-            body: formData,
-          });
-
-          if (!retryResponse.ok) {
-            const retryErrorText = await retryResponse.text();
-            return {
-              data: null,
-              error: {
-                message: `Failed to upload waiver after token refresh: ${retryResponse.status} ${retryErrorText}`,
-                status: retryResponse.status,
-                type: retryResponse.status === 401 ? 'auth' : 'unknown',
-                userMessage: retryResponse.status === 401
-                  ? "Session expired. Please log in again."
-                  : "Failed to upload waiver. Please try again."
-              }
-            };
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (refreshError) {
-          return {
-            data: null,
-            error: {
-              message: "Token refresh failed",
-              status: 401,
-              type: 'auth',
-              userMessage: "Session expired. Please log in again."
-            }
-          };
-        }
-      }
-
-      const errorText = await response.text();
-      const errorType: APIErrorType =
-        response.status === 413 ? 'payload_too_large' :
-        response.status === 400 ? 'validation' :
-        response.status >= 500 ? 'server' :
-        'unknown';
-
-      const userMessage =
-        response.status === 413 ? "File size too large. Maximum 10MB allowed." :
-        response.status === 400 ? "Invalid file format. Please upload a valid document." :
-        response.status >= 500 ? "Server error. Please try again later." :
-        "Failed to upload waiver. Please try again.";
-
-      return {
-        data: null,
-        error: {
-          message: `Failed to upload waiver: ${response.status} ${errorText}`,
-          status: response.status,
-          type: errorType,
-          userMessage: userMessage
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to upload waiver:", error);
-    const isNetworkError =
-      error instanceof TypeError &&
-      (error.message.includes('Network request failed') ||
-       error.message.includes('Failed to fetch'));
-
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to upload waiver",
-        status: 500,
-        type: isNetworkError ? 'network' : 'unknown',
-        userMessage: isNetworkError
-          ? "Please check your internet connection and try again."
-          : "An unexpected error occurred. Please try again."
-      }
-    };
-  }
-};
-
-// Get waivers for a specific user
-export const getUserWaivers = async (userId: string): Promise<APIResponse<Waiver[]>> => {
-  try {
-    let jwtToken = await AsyncStorage.getItem("authToken");
-
-    if (!jwtToken) {
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.error("❌ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401,
-            type: 'auth',
-            userMessage: "Session expired. Please log in again."
-          }
-        };
-      }
-    }
-
-    const response = await fetch(`${API_URL}/waivers/user/${userId}`, {
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-      },
-    });
-
-    if (!response.ok) {
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-          const retryResponse = await fetch(`${API_URL}/waivers/user/${userId}`, {
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-            },
-          });
-
-          if (!retryResponse.ok) {
-            const retryErrorText = await retryResponse.text();
-            return {
-              data: null,
-              error: {
-                message: `Failed to fetch waivers after token refresh: ${retryResponse.status} ${retryErrorText}`,
-                status: retryResponse.status,
-                type: retryResponse.status === 401 ? 'auth' : 'unknown',
-                userMessage: retryResponse.status === 401
-                  ? "Session expired. Please log in again."
-                  : "Unable to load waivers. Please try again."
-              }
-            };
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (refreshError) {
-          return {
-            data: null,
-            error: {
-              message: "Token refresh failed",
-              status: 401,
-              type: 'auth',
-              userMessage: "Session expired. Please log in again."
-            }
-          };
-        }
-      }
-
-      const errorText = await response.text();
-      return {
-        data: null,
-        error: {
-          message: `Failed to fetch waivers: ${response.status} ${errorText}`,
-          status: response.status,
-          type: response.status === 404 ? 'not_found' : 'unknown',
-          userMessage: response.status === 404
-            ? "No waivers found."
-            : "Unable to load waivers. Please try again."
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to fetch waivers:", error);
-    const isNetworkError =
-      error instanceof TypeError &&
-      (error.message.includes('Network request failed') ||
-       error.message.includes('Failed to fetch'));
-
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to fetch waivers",
-        status: 500,
-        type: isNetworkError ? 'network' : 'unknown',
-        userMessage: isNetworkError
-          ? "Please check your internet connection and try again."
-          : "An unexpected error occurred. Please try again."
-      }
-    };
-  }
-};
-
-// Get user subsidy usage history
-export const getUserSubsidyUsage = async (): Promise<APIResponse<any>> => {
-  try {
-    let jwtToken = await AsyncStorage.getItem("authToken");
-
-    if (!jwtToken) {
-      try {
-        jwtToken = await refreshBackendJwt();
-      } catch (refreshError) {
-        console.error("❌ Failed to refresh JWT token:", refreshError);
-        return {
-          data: null,
-          error: {
-            message: "Unable to authenticate with backend",
-            status: 401,
-            type: 'auth',
-            userMessage: "Session expired. Please log in again."
-          }
-        };
-      }
-    }
-
-    const response = await fetch(`${API_URL}/subsidies/me/usage`, {
-      headers: {
-        "Authorization": `Bearer ${jwtToken}`,
-      },
-    });
-
-    if (!response.ok) {
-      if (response.status === 401 || response.status === 403) {
-        try {
-          jwtToken = await refreshBackendJwt();
-          const retryResponse = await fetch(`${API_URL}/subsidies/me/usage`, {
-            headers: {
-              "Authorization": `Bearer ${jwtToken}`,
-            },
-          });
-
-          if (!retryResponse.ok) {
-            const retryErrorText = await retryResponse.text();
-            const retryErrorType: APIErrorType = retryResponse.status === 401 ? 'auth' : 'unknown';
-            const retryUserMessage = retryResponse.status === 401
-              ? "Session expired. Please log in again."
-              : "Unable to load subsidy usage history. Please try again.";
-
-            return {
-              data: null,
-              error: {
-                message: `Failed to fetch subsidy usage after token refresh: ${retryResponse.status} ${retryErrorText}`,
-                status: retryResponse.status,
-                type: retryErrorType,
-                userMessage: retryUserMessage
-              }
-            };
-          }
-
-          const retryData = await retryResponse.json();
-          return { data: retryData, error: null };
-        } catch (refreshError) {
-          return {
-            data: null,
-            error: {
-              message: "Token refresh failed",
-              status: 401,
-              type: 'auth',
-              userMessage: "Session expired. Please log in again."
-            }
-          };
-        }
-      }
-
-      const errorText = await response.text();
-      return {
-        data: null,
-        error: {
-          message: `Failed to fetch subsidy usage: ${response.status} ${errorText}`,
-          status: response.status,
-          type: 'unknown',
-          userMessage: "Unable to load subsidy usage history. Please try again."
-        }
-      };
-    }
-
-    const data = await response.json();
-    return { data, error: null };
-  } catch (error) {
-    console.error("❌ Failed to fetch subsidy usage:", error);
-    return {
-      data: null,
-      error: {
-        message: (error as Error).message || "Failed to fetch subsidy usage",
-        status: 500,
-        type: 'network',
-        userMessage: "Network error. Please check your connection and try again."
-      }
-    };
-  }
-}
+// Legacy entry point: re-export all API modules from utils/api/
+export * from "./api/index";
